Add files via upload
this program will receive an integer array filled with values, and we also are given an integer target that must be found using two values from the array and then return the indices of the two numbers if they add up to equal the target.  you may assume that each input would have only exactly one solution, and you may not use the same element twice.