Add files via upload
given an int array, nums, sorted in non-decreasing order, remove the duplicates in place such that each unique element appears only once. the relative order of the elements should be kept the same. then return the number of unique elements in nums. note that is question was very weird, as it really wants the length / index of the unique elements in the array, so for the ex {1,1,2} the index would be 2 as one is the first index, and two is the second.