Add files via upload
given an int array nums, and an integer, val, remove all occurrences of val in nums in place. The order of the elements may be changed. then return the number of elements in nums that are not equal to val. consider the number of elements which are not equal to val to be k. to get accepted, change the array nums such that the first elements of nums contain the elements which are not equal to val, and the remaining elements of nums are not important as well as size of nums